+++ archive/c/c-plus-plus/longest-common-subsequence.cpp
