+++ archive/p/python/longest_common_subsequence.py
